test(publisher): separate snapshot page-index race coverage
